+++ src/parse.rs
@@ -76,16 +76,6 @@ where
     }
 }
 
-pub fn pat_to_expr<L: Language>(pat: Pattern<L>) -> Result<RecExpr<L>> {
-    match pat {
-        Pattern::Wildcard(w, _) => Err(ParseError(format!(
-            "Found wildcard {:?} instead of expr term",
-            w
-        ))),
-        Pattern::Expr(e) => Ok(e.map_children_result(pat_to_expr)?.into()),
-    }
-}
-
 /// A trait to parse stuff from a `Language`.
 ///
 /// This is blanket-impled for any `Language` whose domains all
@@ -104,7 +94,7 @@ where
 
     fn parse_expr(s: &str) -> Result<RecExpr<Self>> {
         let pat = Self::parse_pattern(s)?;
-        pat_to_expr(pat)
+        pat.to_expr().map_err(ParseError)
     }
 }
